Move source to 'src', add 'py.typed' and mypy checking
