+++ noxfile.py
@@ -1,6 +1,6 @@
 import nox
 
-SOURCE_FILES = ("noxfile.py", "truststore.py", "test_truststore.py")
+SOURCE_FILES = ("noxfile.py", "src/", "test_truststore.py")
 
 
 @nox.session
@@ -14,13 +14,14 @@ def format(session):
 
 @nox.session
 def lint(session):
-    session.install("black", "isort", "flake8")
+    session.install("black", "isort", "flake8", "mypy", "types-certifi")
     session.run("flake8", "--ignore=E501,W503", *SOURCE_FILES)
     session.run("black", "--check", *SOURCE_FILES)
     session.run("isort", "--check", "--profile=black", *SOURCE_FILES)
+    session.run("mypy", "--strict", "--show-error-codes", "src/")
 
 
-@nox.session
+@nox.session(python="3.10")
 def test(session):
-    session.install("-r", "dev-requirements.txt")
+    session.install("-rdev-requirements.txt", ".")
     session.run("pytest", *(session.posargs or ("test_truststore.py",)))
